Few changes don in malawi dts shipment and api service delete option check

+++ application/services/Shipments.php
@@ -2879,12 +2879,14 @@ class Application_Service_Shipments
             $db = Zend_Db_Table_Abstract::getDefaultAdapter();
             $responseTable = array('response_result_dbs', 'response_result_dts', 'response_result_eid', 'response_result_recency', 'response_result_tb', 'response_result_vl');
             foreach ($responseTable as $response) {
-                //   $shipment = $db->fetchRow($db->select()->from($response, array('shipment_map_id'))->where('shipment_map_id =?', $mapId)->where('sample_id =?', $sId));
-                //   if ($shipment) {
-                $db->delete($response, "shipment_map_id = " . $mapId);
-                //   }
+                $shipment = $db->fetchRow($db->select()->from($response, array('shipment_map_id'))->where('shipment_map_id =?', $mapId)->where('sample_id =?', $sId));
+                if ($shipment) {
+                    $db->query("SET FOREIGN_KEY_CHECKS = 0;"); // Disable foreign key checks
+                    $db->delete($response, "shipment_map_id = " . $mapId);
+                }
             }
             return  $db->delete('shipment_participant_map', "map_id = " . $mapId);
+            $db->query("SET FOREIGN_KEY_CHECKS = 1;"); // Enable foreign key checks
         } catch (Exception $e) {
             error_log($e->getMessage());
             return "Unable to delete. Please try again later or contact system admin for help";
